+++ cmd/utils.go
@@ -2,6 +2,8 @@ package cmd
 
 import (
 	"fmt"
+	"regexp"
+	"strings"
 	"time"
 
 	"github.com/hashicorp/hcl/v2/hclsimple"
@@ -10,6 +12,7 @@ import (
 
 	tfcw "github.com/mvisonneau/tfcw/lib/client"
 	"github.com/mvisonneau/tfcw/lib/schemas"
+	"github.com/mvisonneau/tfcw/lib/terraform"
 
 	log "github.com/sirupsen/logrus"
 )
@@ -28,28 +31,30 @@ func configure(ctx *cli.Context) (c *tfcw.Client, cfg *schemas.Config, err error
 		return
 	}
 
-	cfg = &schemas.Config{}
-	log.Debugf("Using config file at %s", ctx.GlobalString("config-path"))
-	err = hclsimple.DecodeFile(ctx.GlobalString("config-path"), nil, cfg)
-	if err != nil {
-		return c, cfg, fmt.Errorf("tfcw config/hcl: %s", err.Error())
+	cfg = &schemas.Config{
+		Runtime: schemas.Runtime{
+			WorkingDir: ctx.GlobalString("working-dir"),
+		},
 	}
 
-	clientConfig := &tfcw.Config{
-		Config: cfg,
+	tfcwConfigFile := computeConfigFilePath(cfg.Runtime.WorkingDir, ctx.GlobalString("config-file"))
+	log.Debugf("Using config file at %s", tfcwConfigFile)
+	err = hclsimple.DecodeFile(tfcwConfigFile, nil, cfg)
+	if err != nil {
+		return c, cfg, fmt.Errorf("tfcw config/hcl: %s", err.Error())
 	}
 
 	// We do not need TFC access to render variables locally
 	switch ctx.Command.Name {
 	case "local":
-		clientConfig.Runtime.TFE.Disabled = true
+		cfg.Runtime.TFC.Disabled = true
 	default:
-		clientConfig.Runtime.TFE.Disabled = false
-		clientConfig.Runtime.TFE.Address = ctx.GlobalString("tfc-address")
-		clientConfig.Runtime.TFE.Token = ctx.GlobalString("tfc-token")
+		if err = computeRuntimeConfigurationForTFC(cfg, ctx); err != nil {
+			return
+		}
 	}
 
-	c, err = tfcw.NewClient(clientConfig)
+	c, err = tfcw.NewClient(cfg)
 	return
 }
 
@@ -67,4 +72,145 @@ func ExecWrapper(f func(ctx *cli.Context) (int, error)) func(*cli.Context) error
 	return func(ctx *cli.Context) error {
 		return exit(f(ctx))
 	}
+}
+
+func computeConfigFilePath(workingDir, configFile string) string {
+	return strings.NewReplacer("<working-dir>", workingDir).Replace(configFile)
+}
+
+func computeRuntimeConfigurationForTFC(cfg *schemas.Config, ctx *cli.Context) (err error) {
+	if cfg.TFC == nil {
+		cfg.TFC = &schemas.TFC{}
+	}
+
+	if cfg.TFC.Workspace == nil {
+		cfg.TFC.Workspace = &schemas.Workspace{}
+	}
+
+	// Address
+	cfg.Runtime.TFC.Address, err = computeRuntimeTFCAddress(cfg.Runtime.WorkingDir, ctx.GlobalString("address"), cfg.TFC.Address)
+	if err != nil {
+		return
+	}
+
+	// Token
+	cfg.Runtime.TFC.Token, err = computeRuntimeTFCToken(cfg.Runtime.WorkingDir, ctx.GlobalString("token"), cfg.TFC.Token)
+	if err != nil {
+		return
+	}
+
+	// Organization
+	cfg.Runtime.TFC.Organization, err = computeRuntimeTFCOrganization(cfg.Runtime.WorkingDir, ctx.GlobalString("organization"), cfg.TFC.Organization)
+	if err != nil {
+		return
+	}
+
+	// Workspace
+	cfg.Runtime.TFC.Workspace, err = computeRuntimeTFCWorkspace(cfg.Runtime.WorkingDir, ctx.GlobalString("workspace"), cfg.TFC.Workspace.Name)
+	return
+}
+
+func computeRuntimeTFCAddress(workingDir, flagValue string, tfcwValue *string) (string, error) {
+	if flagValue != "" {
+		log.Debugf("Using TFC address '%s' from CLI flag (or env variable)", returnHTTPSPrefixedURL(flagValue))
+		return returnHTTPSPrefixedURL(flagValue), nil
+	}
+
+	if tfcwValue != nil {
+		log.Debugf("Using TFC address '%s' from TFCW config", returnHTTPSPrefixedURL(*tfcwValue))
+		return returnHTTPSPrefixedURL(*tfcwValue), nil
+	}
+
+	rbc, err := terraform.GetRemoteBackendConfig(workingDir)
+	if err != nil {
+		return "", err
+	}
+
+	if rbc.Hostname != "" {
+		log.Debugf("Using TFC address '%s' from Terraform remote backend configuration", returnHTTPSPrefixedURL(rbc.Hostname))
+		return returnHTTPSPrefixedURL(rbc.Hostname), nil
+	}
+
+	log.Debug("Using default TFC address 'https://app.terraform.io'")
+	return "https://app.terraform.io", nil
+}
+
+func computeRuntimeTFCToken(workingDir, flagValue string, tfcwValue *string) (string, error) {
+	if flagValue != "" {
+		log.Debug("Using TFC token '***' from CLI flag (or env variable)")
+		return flagValue, nil
+	}
+
+	if tfcwValue != nil {
+		log.Debug("Using TFC token '***' from TFCW config")
+		return *tfcwValue, nil
+	}
+
+	rbc, err := terraform.GetRemoteBackendConfig(workingDir)
+	if err != nil {
+		return "", err
+	}
+
+	if rbc.Token != "" {
+		log.Debug("Using TFC token '***' from Terraform remote backend configuration")
+		return rbc.Token, nil
+	}
+
+	return "", nil
+}
+
+func computeRuntimeTFCOrganization(workingDir, flagValue string, tfcwValue *string) (string, error) {
+	if flagValue != "" {
+		log.Debugf("Using TFC organization '%s' from CLI flag (or env variable)", flagValue)
+		return flagValue, nil
+	}
+
+	if tfcwValue != nil {
+		log.Debugf("Using TFC organization '%s' from TFCW config", *tfcwValue)
+		return *tfcwValue, nil
+	}
+
+	rbc, err := terraform.GetRemoteBackendConfig(workingDir)
+	if err != nil {
+		return "", err
+	}
+
+	if rbc.Organization != "" {
+		log.Debugf("Using TFC organization '%s' from Terraform remote backend configuration", rbc.Organization)
+		return rbc.Organization, nil
+	}
+
+	return "", nil
+}
+
+func computeRuntimeTFCWorkspace(workingDir, flagValue string, tfcwValue *string) (string, error) {
+	if flagValue != "" {
+		log.Debugf("Using TFC workspace '%s' from CLI flag (or env variable)", flagValue)
+		return flagValue, nil
+	}
+
+	if tfcwValue != nil {
+		log.Debugf("Using TFC workspace '%s' from TFCW config", *tfcwValue)
+		return *tfcwValue, nil
+	}
+
+	rbc, err := terraform.GetRemoteBackendConfig(workingDir)
+	if err != nil {
+		return "", err
+	}
+
+	if rbc.Workspace != "" {
+		log.Debugf("Using TFC workspace '%s' from Terraform remote backend configuration", rbc.Workspace)
+		return rbc.Workspace, nil
+	}
+
+	return "", nil
+}
+
+func returnHTTPSPrefixedURL(url string) string {
+	re := regexp.MustCompile(`^(http|https)://`)
+	if re.MatchString(url) {
+		return url
+	}
+	return fmt.Sprintf("https://%s", url)
 }