+++ lib/schemas/config.go
@@ -7,4 +7,19 @@ type Config struct {
 	Defaults             *Defaults `hcl:"defaults,block"`
 	TerraformVariables   Variables `hcl:"tfvar,block"`
 	EnvironmentVariables Variables `hcl:"envvar,block"`
+
+	Runtime Runtime
+}
+
+// Runtime is a struct used by the client in order
+// to store values configured at runtime
+type Runtime struct {
+	WorkingDir string
+	TFC        struct {
+		Disabled     bool
+		Address      string
+		Token        string
+		Organization string
+		Workspace    string
+	}
 }