+++ lib/schemas/tfc.go
@@ -2,8 +2,10 @@ package schemas
 
 // TFC handles Terraform Cloud related configuration
 type TFC struct {
-	Organization string    `hcl:"organization"`
-	Workspace    Workspace `hcl:"workspace,block"`
+	Address      *string    `hcl:"address"`
+	Token        *string    `hcl:"token"`
+	Organization *string    `hcl:"organization"`
+	Workspace    *Workspace `hcl:"workspace,block"`
 
 	WorkspaceAutoCreate     *bool `hcl:"workspace-auto-create"`
 	PurgeUnmanagedVariables *bool `hcl:"purge-unmanaged-variables"`
@@ -11,8 +13,7 @@ type TFC struct {
 
 // Workspace is used to refer to and configure the workspace
 type Workspace struct {
-	Name string `hcl:"name"`
-
+	Name             *string `hcl:"name"`
 	Operations       *bool   `hcl:"operations"`
 	AutoApply        *bool   `hcl:"auto-apply"`
 	TerraformVersion *string `hcl:"terraform-version"`