+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [0ver](https://0ver.org).
 
 ## [Unreleased]
 
+## Added
+
+- Organization and workspace configuration can now either be set directly through respective flags `--organization` / `--workspace`
+- TFC configuration (address, token, organization & workspace)will default to what is configured as a remote backend in the Terraform configuration
+
 ## [0.0.4] - 2020-04-01
 
 ### Added